requirements.txt
pyyaml
wheel
setuptools
twine
pdoc3
pytest-cov
Click
flake8
